Bootstrap the cornice webapp.
It's pluggable so it's possible to change the backend to use directly from the
configuration file. This is made to allow easy configuration in the future.

The default implementation will probably be the RAW SQL since that's a small
project. In the future, we could do some more complex stuff, like using caching
strategies.